requirements.txt
SQLAlchemy==2.0.27
alembic==1.13.1
psycopg==3.1.18
psycopg2-binary
faker
factory-boy